move some of vec into headeronly in preparation for Half.h
ghstack-source-id: ec6f15fa42bc9a54f4eb5c02b854cb507cd6b6e5
Pull Request resolved: https://github.com/pytorch/pytorch/pull/158976
