+++ torch/headeronly/CMakeLists.txt
@@ -20,6 +20,7 @@ configure_file(
 
 file(GLOB HEADERONLY_HEADERS
     *.h
+    cpu/**/*.h
     macros/*.h
     util/*.h
 )
